refactor(services): reorganize service file structure
- Move cookie_client.dart → auth/cookie_client.dart
- Move scu_exceptions.dart → auth/scu_exceptions.dart
- Move ccyl_oauth_service.dart → auth/ccyl_oauth_service.dart
- Move ccyl_service.dart → ccyl/ccyl_service.dart
- Move balance_query_service.dart → balance/balance_query_service.dart
- Delete empty scu_api/ and exceptions/ directories
- Delete old barrel file scu_api_service.dart
