+++ lib/pages/campus/ccyl/credit_list_page.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:bugaoshan/injection/injector.dart';
 import 'package:bugaoshan/l10n/app_localizations.dart';
 import 'package:bugaoshan/providers/ccyl_provider.dart';
-import 'package:bugaoshan/services/ccyl_service.dart';
+import 'package:bugaoshan/services/ccyl/ccyl_service.dart';
 import 'package:bugaoshan/widgets/common/error_widgets.dart';
 
 class CreditListPage extends StatefulWidget {
