+++ lib/pages/campus/fitness_test/fitness_test_page.dart
@@ -6,8 +6,8 @@ import 'package:bugaoshan/injection/injector.dart';
 import 'package:bugaoshan/l10n/app_localizations.dart';
 import 'package:bugaoshan/providers/scu_auth_provider.dart';
 import 'package:bugaoshan/services/auth/fitness_auth.dart';
-import 'package:bugaoshan/services/exceptions/scu_exceptions.dart';
-import 'package:bugaoshan/services/scu_api/cookie_client.dart';
+import 'package:bugaoshan/services/auth/scu_exceptions.dart';
+import 'package:bugaoshan/services/auth/cookie_client.dart';
 import 'package:bugaoshan/utils/constants.dart';
 import 'package:bugaoshan/widgets/common/loading_widgets.dart';
 import 'package:bugaoshan/widgets/common/login_required_widget.dart';
