+++ lib/services/auth/cookie_client.dart
