--- lib/services/scu_api_service.dart
@@ -1,6 +1,0 @@
-// Barrel export — 保持旧 import 路径兼容。
-// 已迁移到三层架构，此文件仅导出新路径。
-export 'package:bugaoshan/services/scu_api/cookie_client.dart';
-export 'package:bugaoshan/services/auth/scu_auth.dart'
-    show kZhjwBase, CaptchaResult;
-export 'package:bugaoshan/services/exceptions/scu_exceptions.dart';
